moved the notes page to activity_notes to allocate the activity_main for the home page, made the notes button go to the notes page on click
