+++ app/src/main/AndroidManifest.xml
@@ -28,10 +28,16 @@
             android:name=".Login"
             android:exported="false" />
         <activity
-            android:name=".navdrawer"
-            android:exported="false"
-            android:label="@string/title_activity_navdrawer"
-            android:theme="@style/Theme.Inkwell.NoActionBar" />
+            android:name=".NotesActivity"
+            android:exported="true"
+            android:label="@string/app_name"
+            android:theme="@style/Theme.Inkwell.NoActionBar">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity
             android:name=".MainActivity"
             android:exported="true"
